SpamKeeperBot/requirements.txt
python-telegram-bot
numpy
transformers
torch
"python-telegram-bot[job-queue]"
